+++ src/Libxc.jl
@@ -1,15 +1,18 @@
 module Libxc
 using Libxc_jll: libxc
+import Libxc_GPU_jll
 using Requires
 
 include("gen/common.jl")
 include("gen/api.jl")
 include("Functional.jl")
-include("functional_gpu.jl")
 include("evaluate.jl")
 
 const libxc_version = VersionNumber(XC_VERSION)
 const libxc_doi = unsafe_string(Libxc.xc_reference_doi())
+
+"""Is the CUDA version of libxc available on this platform"""
+has_cuda() = isdefined(Libxc_GPU_jll, :libxc)
 
 """Return the list of available libxc functionals as strings"""
 function available_functionals()
@@ -31,7 +34,11 @@ export is_lda, is_gga, is_mgga, is_hybrid, is_vv10, is_range_separated, is_globa
 export needs_laplacian
 
 function __init__()
-    @require CUDA="052768ef-5323-5732-b1bb-66c8b64840ba" include("evaluate_gpu.jl")
+    @require CUDA="052768ef-5323-5732-b1bb-66c8b64840ba" begin
+        @static if isdefined(Libxc_GPU_jll, :libxc)
+            include("evaluate_gpu.jl")
+        end
+    end
 end
 
 end # module
